fix mapper. fix dto. fix update service. fix create service

+++ src/main/java/br/com/alelo/consumer/consumerpat/respository/ConsumerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface ConsumerRepository extends JpaRepository<Consumer, Integer> {
+public interface ConsumerRepository extends JpaRepository<Consumer, Long> {
 
     @Query(nativeQuery = true, value = "select * from Consumer")
     List<Consumer> getAllConsumersList();
